Big rewrite for perpendicular variant

+++ src/Topics/Physics.HomongenousParticle.Logic/PhysicsServices/IPhysicsService.cs
@@ -6,9 +6,9 @@ namespace Physics.HomongenousParticle.Logic.PhysicsServices
 {
     interface IPhysicsService
     {
-        float ComputeX(float seconds);
+        double ComputeX(float seconds);
 
-        float ComputeY(float seconds);
+        double ComputeY(float seconds);
 
         float MaxT { get; }
     }
